+++ packages/e2e-test-utils/src/set-option.js
@@ -9,42 +9,24 @@ import { pressKeyWithModifier } from './press-key-with-modifier';
 /**
  * Sets a site option, from the options-general admin page.
  *
- * @param {string} setting   The option, used to get the option by id.
- * @param {string} value     The value to set the option to.
- * @param {string} adminPage The url of the admin page to visit. Default: options-general.php.
+ * @param {string} setting The option, used to get the option by id.
+ * @param {string} value   The value to set the option to.
  *
  */
-export async function setOption(
-	setting,
-	value,
-	adminPage = 'options-general.php'
-) {
+export async function setOption( setting, value ) {
 	await switchUserToAdmin();
-	await visitAdminPage( adminPage );
+	await visitAdminPage( 'options.php' );
 
-	const optionType = await page.$eval(
-		`#${ setting }`,
-		( element ) => element.type
-	);
-	if ( optionType === 'checkbox' ) {
-		const isChecked = await page.$eval(
-			`#${ setting }`,
-			( element ) => element.checked === true
-		);
-		if ( ( value && ! isChecked ) || ( ! value && isChecked ) ) {
-			await page.click( `#${ setting }` );
-		}
-	} else if ( optionType === 'select-one' ) {
-		await page.select( `#${ setting }`, value );
-	} else {
-		// Update other options types.
-		await page.focus( `#${ setting }` );
-		await pressKeyWithModifier( 'primary', 'a' );
-		await page.type( `#${ setting }`, value );
+	if ( typeof value === 'boolean' ) {
+		value = value ? '1' : '0';
 	}
 
+	await page.focus( `#${ setting }` );
+	await pressKeyWithModifier( 'primary', 'a' );
+	await page.type( `#${ setting }`, value );
+
 	await Promise.all( [
-		page.click( '#submit' ),
+		page.click( '#Update' ),
 		page.waitForNavigation( { waitUntil: 'networkidle0' } ),
 	] );
 	await switchUserToTest();